+++ public/admin/_app/immutable/nodes/24.c7M2ooau.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/Cm_fJlcz.js";export{e as component};